1.2 src-code update
yay

+++ PR-DOS/Module1.vb
@@ -2,7 +2,50 @@
     'The sub "Main" is the normal version of PR-DOS and is launched by default.
     'While the sub "graphical" is the GPR-DOS, which can be launched from normal mode using the "graphical" command.
     Sub Main()
+#Disable Warning BC42104
         Console.Clear()
+        If My.Computer.FileSystem.FileExists("C:\prdos_settings\bricked.txt") = True Then
+brick:
+            Console.Clear()
+            Console.WriteLine("YOUR PR-DOS SYSTEM IS BRICKED! PLEASE REINSTALL PR-DOS BY DELETING THE FOLDER")
+            Console.WriteLine("prdos_settings AND RE-DOWNLOADING PR-DOS!")
+            Console.WriteLine("Press ENTER to shutdown...")
+            Dim wow = Console.ReadLine()
+            If wow = "pardon" Then
+                Call secret()
+            End If
+            End
+        End If
+        If My.Computer.FileSystem.FileExists("C:\nrun_clever.txt") = True Then
+
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_original.txt") = True Then
+
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_besty.txt") = True Then
+
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_xtreme.txt") = True Then
+
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_nelpower.txt") = True Then
+
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_server.txt") = True Then
+
+        Else
+            Console.WriteLine("A N-rUn BASE IS NEEDED FOR PR-DOS TO RUN! PLEASE GET A N-rUn BASE AND THEN RUN PR-DOS AGAIN.")
+            Console.WriteLine("Press ENTER to shutdown...")
+            Console.ReadLine()
+            End
+        End If
+        If My.Computer.FileSystem.FileExists("C:\prdos_settings\glaunch_prohibit.txt") = True Then
+            GoTo prevent
+        End If
+        If My.Computer.FileSystem.FileExists("C:\prdos_settings\glaunch.txt") = True Then
+            Call graphical()
+        End If
+prevent:
+        Dim news As String
+        If My.Computer.FileSystem.FileExists("C:\prdos_settings\nodown.txt") = True Then
+            news = "Prevented from downloading news."
+            GoTo fuck
+        End If
         Dim remoteUri As String = "http://nelinka.6te.net/prdos/news.txt"
         Dim fileName As String = "C:\news.txt"
 
@@ -11,11 +54,54 @@
             client.Credentials = New System.Net.NetworkCredential
             client.DownloadFile(remoteUri, fileName)
         End Using
-        Dim news As String
         news = My.Computer.FileSystem.ReadAllText("C:\news.txt")
-        Console.WriteLine("PR-DOS")
+fuck:
+        Dim nrun As String
+        Dim nrun_number As String
+        If My.Computer.FileSystem.FileExists("C:\nrun_clever.txt") = True Then
+            nrun_number = My.Computer.FileSystem.ReadAllText("C:\nrun_clever.txt")
+            nrun = "N-rUn Clever "
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_original.txt") = True Then
+            nrun_number = My.Computer.FileSystem.ReadAllText("C:\nrun_original.txt")
+            nrun = "N-rUn Original "
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_besty.txt") = True Then
+            nrun_number = My.Computer.FileSystem.ReadAllText("C:\nrun_besty.txt")
+            nrun = "N-rUn BESTy "
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_xtreme.txt") = True Then
+            nrun_number = My.Computer.FileSystem.ReadAllText("C:\nrun_xtreme.txt")
+            nrun = "N-rUn Xtreme "
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_nelpower.txt") = True Then
+            nrun_number = My.Computer.FileSystem.ReadAllText("C:\nrun_nelpower.txt")
+            nrun = "N-rUn NELPOWER "
+        ElseIf My.Computer.FileSystem.FileExists("C:\\nrun_server.txt") = True Then
+            nrun_number = My.Computer.FileSystem.ReadAllText("C:\nrun_server.txt")
+            nrun = "N-rUn Server_ "
+        End If
+        Dim modified As Boolean
+        modified = False
+        If nrun_number = "MODIFIED" Then
+            modified = True
+        End If
+secretway:
+        Console.Clear()
+        Console.ForegroundColor = ConsoleColor.Red
+        Console.WriteLine("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
+        Console.WriteLine("X  OOO   OOO        OOOO    OOO    OOOO   X")
+        Console.WriteLine("X  O  O  O  O       O   O  O   O  O       X")
+        Console.WriteLine("X  OOO   OOO   ---  O   O  O   O   OOOO   X")
+        Console.WriteLine("X  O     O  O       O   O  O   O       O  X")
+        Console.WriteLine("X  O     O  O       OOOO    OOO    OOOO   X")
+        Console.WriteLine("XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX")
+        Console.ForegroundColor = ConsoleColor.White
+        Console.WriteLine("Normal MODE")
+        Console.Write("Time: ")
+        Console.ForegroundColor = ConsoleColor.Yellow
+        Console.WriteLine(Now.ToShortTimeString)
+        Console.ForegroundColor = ConsoleColor.White
         Console.WriteLine("----------------------")
-        Console.WriteLine("Bringing DOS back, by Nelinka. Ver. 1.1")
+        Console.WriteLine("Bringing DOS back, by Nelinka. Ver. 1.2")
+        Console.Write(nrun)
+        Console.WriteLine(nrun_number)
         Console.WriteLine("-----------------------------------------------")
         Console.WriteLine("TIP: If you want PR-DOS to look real, press F11! (or FN+F11)")
         Console.WriteLine("Are you lost? Enter 'help' in the CMD to get a list of commands!")
@@ -29,6 +115,9 @@ beginning:
         If theinput = "shutdown" Then
             Console.WriteLine("shutdown>Ready to close PR-DOS! Press ENTER to close PR-DOS.")
             Console.ReadLine()
+            If My.Computer.FileSystem.FileExists("C:\prdos_settings\glaunch_prohibit.txt") = True Then
+                My.Computer.FileSystem.DeleteFile("C:\prdos_settings\glaunch_prohibit.txt")
+            End If
             End
         ElseIf theinput = "clear" Then
             Console.Clear()
@@ -44,27 +133,59 @@ beginning:
             Console.WriteLine("help>run = Runs a PR-DOS program you desire.")
             Console.WriteLine("help>delete = Deletes a file you desire.")
             Console.WriteLine("help>rename = Renames a file you desire.")
+            Console.WriteLine("help>dir = Opens a directory you desire.")
+            Console.WriteLine("help>createtxt = Creates a TXT file where you want, including what you want.")
+            Console.WriteLine("help>server = Helps manage a server running a N-rUn Server_ base.")
             GoTo beginning
         ElseIf theinput = "restart" Then
+            If My.Computer.FileSystem.FileExists("C:\prdos_settings\glaunch_prohibit.txt") = True Then
+                My.Computer.FileSystem.DeleteFile("C:\prdos_settings\glaunch_prohibit.txt")
+            End If
             Call Main()
         ElseIf theinput = "about" Then
             Console.WriteLine("about>PR-DOS made by Nelinka.")
-            Console.WriteLine("about>Ver. 1.1 -- CHANGELOG")
+            Console.WriteLine("about>Ver. 1.2 -- CHANGELOG")
             Console.WriteLine("about>-----------------------------------")
-            Console.WriteLine("about>Added commands:")
-            Console.WriteLine("about>'graphical', 'settings', 'run', 'delete' and 'rename'.")
-            Console.WriteLine("about>Renamed commands:")
-            Console.WriteLine("about>'close' to 'shutdown' and 'list' to 'help'.")
-            Console.WriteLine("about>Also added preparations for GPR-DOS, it's not working yet but the command for it is 'graphical'.")
+            Console.WriteLine("about>Added a setting.")
+            Console.WriteLine("about>Added a cool title with some help! Thanks!")
+            Console.WriteLine("about>Added commands 'dir', 'createtxt' and 'server'.")
+            Console.WriteLine("about>Implemented N-rUn bases.")
+            Console.WriteLine("about>Implemented a brick system.")
+            Console.WriteLine("about>Added double-check to the 'delete' command.")
+            Console.WriteLine("about>Added the clock to normal mode too. :D")
+            Console.WriteLine("about>You can now MODIFY your N-rUn Base! :D In the next update or dev-release, it will be good for something!")
+            Console.WriteLine("about>Also you can now prevent the news from downloading.")
+            Console.WriteLine("about>Bug fixes.")
             GoTo beginning
         ElseIf theinput = "graphical" Then
             Call graphical()
         ElseIf theinput = "settings" Then
-            Console.WriteLine("settings>There are no settings for PR-DOS/GPR-DOS! Yet..")
+            If My.Computer.FileSystem.DirectoryExists("C:\prdos_settings") = False Then
+                My.Computer.FileSystem.CreateDirectory("C:\prdos_settings")
+            End If
+            Console.WriteLine("settings>Do you want to start GPR-DOS on launch? (y/n)")
+            Dim selection = Console.ReadLine()
+            If selection = "y" Then
+                My.Computer.FileSystem.WriteAllText("C:\prdos_settings\glaunch.txt", "yes", False)
+                Console.WriteLine("settings>Saved this setting.")
+            ElseIf selection = "n" Then
+                If My.Computer.FileSystem.FileExists("C:\prdos_settings\glaunch_prohibit.txt") = True Then
+                    My.Computer.FileSystem.DeleteFile("C:\prdos_settings\glaunch_prohibit.txt")
+                End If
+                If My.Computer.FileSystem.FileExists("C:\prdos_settings\glaunch.txt") = True Then
+                    My.Computer.FileSystem.DeleteFile("C:\prdos_settings\glaunch.txt")
+                    Console.WriteLine("settings>Deleted the setting.")
+                End If
+            End If
+            Console.WriteLine("settings>There are no more settings for PR-DOS/GPR-DOS! Yet..")
             GoTo beginning
         ElseIf theinput = "run" Then
             Console.WriteLine("run>Please enter the program path:")
             Dim filepath = Console.ReadLine()
+            If My.Computer.FileSystem.FileExists(filepath) = False Then
+                Console.WriteLine("run>Error! Please check that the file in the filepath you entered exists!")
+                GoTo beginning
+            End If
             Dim program = My.Computer.FileSystem.ReadAllText(filepath)
             Console.WriteLine(program)
             GoTo beginning
@@ -75,8 +196,13 @@ beginning:
                 Console.WriteLine("delete>Error! Please check if the file exists in the path you entered.")
                 GoTo beginning
             End If
-            My.Computer.FileSystem.DeleteFile(filepath)
-            Console.WriteLine("delete>Deleted ", filepath)
+            Console.WriteLine("delete>Are you sure? (Y/N)")
+            Dim selection = Console.ReadLine()
+            If selection = "y" Then
+                My.Computer.FileSystem.DeleteFile(filepath)
+                Console.WriteLine("delete>Deleted the file in the path you entered.")
+                GoTo beginning
+            End If
             GoTo beginning
         ElseIf theinput = "rename" Then
             Console.WriteLine("rename>Please enter the file path:")
@@ -88,8 +214,37 @@ beginning:
             Console.WriteLine("rename>Please enter what you wanna rename the file to:")
             Dim name = Console.ReadLine()
             My.Computer.FileSystem.RenameFile(filepath, name)
-            Console.WriteLine("rename>Successfully renamed ", filepath, " to ", name)
+            Console.WriteLine("rename>Successfully renamed the file in the path you entered.")
             GoTo beginning
+        ElseIf theinput = "dir" Then
+            Console.WriteLine("dir>Please enter the directory path:")
+            Dim filepath = Console.ReadLine()
+            If My.Computer.FileSystem.DirectoryExists(filepath) = False Then
+                Console.WriteLine("dir>Error! Please check that the directory in the path you entered exists!")
+                GoTo beginning
+            End If
+            Process.Start(filepath)
+            Console.WriteLine("dir>Opened the directory in the path you entered.")
+            GoTo beginning
+        ElseIf theinput = "createtxt" Then
+            Console.WriteLine("createtxt>Please enter the filepath:")
+            Dim filepath = Console.ReadLine()
+            Console.WriteLine("createtxt>Please enter the text that will the TXT contain:")
+            Dim text = Console.ReadLine()
+            My.Computer.FileSystem.WriteAllText(filepath, text, False)
+            Console.WriteLine("createtxt>Created a TXT file.")
+            GoTo beginning
+        ElseIf theinput = "server" Then
+            If nrun = "N-rUn Server_ " Then
+                Console.WriteLine("server>This command is preparing, but thanks for choosing the N-rUn Server_!")
+                GoTo beginning
+            Else
+                Console.WriteLine("server>You need the N-rUn Server_ Base in order to use this command!")
+                GoTo beginning
+            End If
+        ElseIf theinput = "brick" Then
+            My.Computer.FileSystem.WriteAllText("C:\prdos_settings\bricked.txt", "bricked", False)
+            GoTo brick
         Else
             Console.WriteLine("PR-DOS>This command does not exist!")
             GoTo beginning
@@ -97,6 +252,20 @@ beginning:
     End Sub
     Sub graphical()
         Console.Clear()
+        Dim nrun As String
+        If My.Computer.FileSystem.FileExists("C:\nrun_clever.txt") = True Then
+            nrun = "/                   N-rUn Clever                                /"
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_original.txt") = True Then
+            nrun = "/                   N-rUn Original                              /"
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_besty.txt") = True Then
+            nrun = "/                   N-rUn BESTy                                 /"
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_xtreme.txt") = True Then
+            nrun = "/                   N-rUn Xtreme                                /"
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_nelpower.txt") = True Then
+            nrun = "/                   N-rUn NELPOWER                              /"
+        ElseIf My.Computer.FileSystem.FileExists("C:\nrun_server.txt") = True Then
+            nrun = "/                   N-rUn Server_                               /"
+        End If
         Console.WriteLine("/////////////////////////////////////////////////////////////////")
         Console.WriteLine("/                                                               /")
         Console.WriteLine("/                   -----------             -----------X        /")
@@ -115,7 +284,7 @@ beginning:
         Console.WriteLine("/          Copyright © Nelinka 2019                             /")
         Console.WriteLine("/          Copyright © Nelinka Software Development 2019        /")
         Console.WriteLine("/                                                               /")
-        Console.WriteLine("/                                                               /")
+        Console.WriteLine(nrun)
         Console.WriteLine("/////////////////////////////////////////////////////////////////")
         Console.Write("Booting up...")
         Threading.Thread.Sleep(5000)
@@ -188,14 +357,20 @@ input2:
                     GoTo input2
                 End If
             ElseIf input2 = "s" Then
+                My.Computer.FileSystem.WriteAllText("C:\prdos_settings\noshut.txt", "nope", False)
                 GoTo shutdown
             ElseIf input2 = "t" Then
                 GoTo desktop_time_clear
             Else
-                GoTo desktop
+                GoTo input2
             End If
+        Else
+            GoTo desktop
         End If
 shutdown:
+        If My.Computer.FileSystem.FileExists("C:\prdos_settings\noshut.txt") = False Then
+            GoTo desktop_time_clear
+        End If
         Console.Clear()
         Console.WriteLine("/////////////////////////////////////////////////////////////////")
         Console.WriteLine("/                                                               /")
@@ -215,11 +390,22 @@ shutdown:
         Console.WriteLine("/          Copyright © Nelinka 2019                             /")
         Console.WriteLine("/          Copyright © Nelinka Software Development 2019        /")
         Console.WriteLine("/                                                               /")
-        Console.WriteLine("/                                                               /")
+        Console.WriteLine(nrun)
         Console.WriteLine("/////////////////////////////////////////////////////////////////")
         Console.Write("Shutting down...")
         Threading.Thread.Sleep(5000)
         Console.WriteLine("Shutted down, press ENTER to go to PR-DOS..")
+        My.Computer.FileSystem.WriteAllText("C:\prdos_settings\glaunch_prohibit.txt", "prohibited", False)
+        Console.ReadLine()
+        If My.Computer.FileSystem.FileExists("C:\prdos_settings\noshut.txt") = True Then
+            My.Computer.FileSystem.DeleteFile("C:\prdos_settings\noshut.txt")
+        End If
+        Call Main()
+    End Sub
+    Sub secret()
+        Console.WriteLine("Oh! Hello there! Welcome to the secret of PR-DOS! OOOOH!")
+        Console.WriteLine("There will be more in the future!")
+        Console.WriteLine("Okay! Bye!")
         Console.ReadLine()
         Call Main()
     End Sub
